refactor(crm): reorganize v2 utils entity-wise + common
Group v2 utils by entity with one shared commonUtil:
- dealUtil gains resolveDeals; contactUtil gains toContactsRecord
- boardUtil absorbs the board selectors and kanban reorder math (kanbanMath merged in, tests folded into boardUtil.test)
- commonUtil holds formatCurrency + reference-data record builders (owner/stage/taskType) and stage ordering; kept pure so util tests load
- getPriorityConfig isolated in priorityUtil (pulls in UI icons)
- delete crmEntityUtils, selectorUtils, kanbanMath
