--- frontend/src/community/crm/v2/utils/crmEntityUtils.ts
@@ -1,54 +1,0 @@
-import {
-  CrmContactEntity,
-  CrmContactRecord,
-  CrmOwnerEntity,
-  CrmOwnerRecord,
-  CrmStageEntity,
-  CrmStageRecord,
-  CrmTaskTypeEntity,
-  CrmTaskTypeRecord
-} from "~community/crm/v2/types/CrmCommonTypes";
-
-export const toStagesRecord = (stages: CrmStageEntity[]): CrmStageRecord => {
-  const stageRecord: CrmStageRecord = {};
-  for (const stage of stages) {
-    if (stage.id != null) {
-      stageRecord[stage.id] = stage;
-    }
-  }
-  return stageRecord;
-};
-
-export const toOwnersRecord = (owners: CrmOwnerEntity[]): CrmOwnerRecord => {
-  const ownerRecord: CrmOwnerRecord = {};
-  for (const owner of owners) {
-    if (owner.employeeId != null) {
-      ownerRecord[owner.employeeId] = owner;
-    }
-  }
-  return ownerRecord;
-};
-
-export const toContactsRecord = (
-  contacts: CrmContactEntity[]
-): CrmContactRecord => {
-  const contactRecord: CrmContactRecord = {};
-  for (const contact of contacts) {
-    if (contact.id != null) {
-      contactRecord[contact.id] = contact;
-    }
-  }
-  return contactRecord;
-};
-
-export const toTaskTypesRecord = (
-  taskTypes: CrmTaskTypeEntity[]
-): CrmTaskTypeRecord => {
-  const taskTypeRecord: CrmTaskTypeRecord = {};
-  for (const taskType of taskTypes) {
-    if (taskType.id != null) {
-      taskTypeRecord[taskType.id] = taskType;
-    }
-  }
-  return taskTypeRecord;
-};
